requirements.txt
SQLAlchemy
python-dotenv
psycopg2-binary
qdrant-client[fastembed]
requests
redis[hiredis]

numpy
rank_bm25
langchain
langchain-openai
langchain-community
transformers
llama_index
pypdf
llama-index-vector-stores-qdrant
llama-index-readers-file
llama-index-embeddings-fastembed
pymupdf
pdf2image
pytesseract
whisper-openai
unstructured
opencv-python
pytesseract
openai-whisper

torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
email-validator
python-jose
passlib
python-multipart
psycopg2-binary
bcrypt

minio
pymupdf
python-magic
python-multipart
python-magic
watchdog

aiobotocore
aiofiles
